+++ packages/calcite-components/src/components/dialog/dialog.e2e.ts
@@ -13,6 +13,7 @@ import {
 } from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
 import { GlobalTestProps, isElementFocused, skipAnimations } from "../../tests/utils";
+import { IDS as PanelIDS } from "../panel/resources";
 import { DialogMessages } from "./assets/dialog/t9n";
 import { CSS, dialogStep, SLOTS } from "./resources";
 
@@ -405,7 +406,7 @@ describe("calcite-dialog", () => {
       await page.waitForChanges();
       expect(await page.find(`calcite-dialog >>> .${CSS.containerOpen}`)).toBeDefined();
 
-      const closeButton = await page.find(`calcite-dialog >>> calcite-panel >>> [data-test="close"]`);
+      const closeButton = await page.find(`calcite-dialog >>> calcite-panel >>> #${PanelIDS.close}`);
       await closeButton.click();
       await page.waitForChanges();
       expect(mockCallBack).toHaveBeenCalledTimes(2);
@@ -729,7 +730,7 @@ describe("calcite-dialog", () => {
     await page.waitForChanges();
     expect(await container.isVisible()).toBe(true);
 
-    const closeButton = await page.find(`calcite-dialog >>> calcite-panel >>> [data-test="close"]`);
+    const closeButton = await page.find(`calcite-dialog >>> calcite-panel >>> #${PanelIDS.close}`);
     await closeButton.click();
     await page.waitForChanges();
     expect(await container.isVisible()).toBe(false);
@@ -1093,5 +1094,22 @@ describe("calcite-dialog", () => {
         },
       },
     );
+
+    it("should not close when slotted panels are closed", async () => {
+      const page = await newE2EPage({
+        html: html`<calcite-dialog open>
+          <calcite-panel closable heading="test"></calcite-panel>
+        </calcite-dialog>`,
+      });
+      await page.waitForChanges();
+
+      const closeButton = await page.find(`calcite-panel >>> #${PanelIDS.close}`);
+
+      await closeButton.click();
+      await page.waitForChanges();
+
+      const dialog = await page.find("calcite-dialog");
+      expect(await dialog.getProperty("open")).toBe(true);
+    });
   });
 });